# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "The body of every error this API answers: the stable code a client\nmatches on (spec §3.4), and a human message nobody should match on. A\nrequest axum rejects before a handler runs (malformed JSON, a missing query\nparameter) gets one too, through\n[`rejections_as_error_bodies`].",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "current_revision": {
            "description": "`REVISION_CONFLICT` only: the resource's revision now, to read again at.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "problems": {
            "description": "`WORKFLOW_VALIDATION_FAILED` only: each problem, the same sentences\n`message` joins.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorCode": {
        "description": "Stable codes clients pattern-match on. Never match on human text.\nSpec §3.4. `Blocked`/`Rejected` are domain outcomes and never appear here.",
        "enum": [
          "HARNESS_START_FAILED",
          "PROCESS_SPAWN_FAILED",
          "PROCESS_TERMINATED",
          "PROCESS_TERMINATION_FAILED",
          "RUNTIME_STOPPING",
          "STORAGE_UNAVAILABLE",
          "STORAGE_MIGRATION_FAILED",
          "STORAGE_CONSTRAINT_VIOLATION",
          "COMMAND_CONFLICT",
          "IDEMPOTENCY_KEY_REQUIRED",
          "INVALID_COMMAND",
          "INVALID_CURSOR",
          "AGENT_AUTH_FAILED",
          "AGENT_UNSUPPORTED_PROFILE",
          "PATH_INVALID",
          "PATH_NOT_FOUND",
          "PATH_NOT_A_DIRECTORY",
          "PATH_ACCESS_DENIED",
          "PATH_ALREADY_EXISTS",
          "PATH_UNAVAILABLE",
          "ORIGIN_REFUSED",
          "HARNESS_UNAVAILABLE",
          "SETTING_NOT_OFFERED",
          "MODE_NOT_ALLOWED",
          "HARNESS_LOCKED",
          "THREAD_BUSY",
          "FORK_POINT_NOT_SUPPORTED",
          "WORKFLOW_FROZEN_IMMUTABLE",
          "WORKFLOW_VALIDATION_FAILED",
          "REVISION_CONFLICT",
          "PROJECT_HAS_THREADS",
          "GRANT_SCOPE",
          "GRANT_INVALID"
        ],
        "type": "string"
      },
...
      "ForkThread": {
        "description": "Forks the thread from its last completed entry (spec §12.9).",
        "properties": {
          "at_entry_id": {
            "$ref": "#/components/schemas/ThreadEntryId"
          },
          "command_id": {
            "description": "The idempotency key (spec §3.2), scoped to the source thread.",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "at_entry_id"
        ],
        "type": "object"
      },
...
      "PlanningThread": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "forked_from_thread": {
            "description": "The thread this one was forked from, if it is a fork (spec §12.9).",
            "type": [
              "string",
              "null"
            ]
          },
          "harness": {
            "description": "The CLI this conversation runs on (`agent::policy`): chosen at\ncreation, changeable until the thread's first operation, then fixed\n(spec §12.6).",
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/ThreadId"
          },
          "project_id": {
            "$ref": "#/components/schemas/ProjectId"
          },
          "removed_at": {
            "description": "Set when this conversation was removed; its history remains readable.",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "title",
          "status",
          "created_at",
          "removed_at",
          "harness",
          "forked_from_thread"
        ],
        "type": "object"
      },
...
      "ProjectId": {
        "format": "uuid",
        "type": "string"
      },
...
      "ThreadEntryId": {
        "format": "uuid",
        "type": "string"
      },
...
      "ThreadId": {
        "format": "uuid",
        "type": "string"
      },
...
      "UpdateThread": {
        "description": "Changes the thread's CLI. Refused once the thread has run a turn, and on\na fork from birth (`HARNESS_LOCKED`).",
        "properties": {
          "command_id": {
            "description": "The idempotency key (spec §3.2), scoped to the thread.",
            "type": "string"
          },
          "harness": {
            "description": "`claude-code` or `codex`.",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "harness"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The local daemon's HTTP API. Every client is on another origin; only origins passed with `--allow-origin` may call it. There is no authentication: the daemon binds to loopback (spec §1, OPEN block on remote access). Errors answered by the API carry `ErrorBody`, whose `code` is stable; a request axum rejects before a handler runs (malformed JSON, a missing parameter) is answered in plain text.",
    "license": {
      "name": ""
    },
    "title": "shadows",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/threads/{id}": {
      "delete": {
        "operationId": "remove_thread",
        "parameters": [
          {
            "description": "The thread",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ThreadId"
            }
          },
          {
            "in": "query",
            "name": "command_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningThread"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "threads"
        ]
      },
      "get": {
        "operationId": "get_thread",
        "parameters": [
          {
            "description": "The thread",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ThreadId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningThread"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "threads"
        ]
      },
      "patch": {
        "operationId": "update_thread",
        "parameters": [
          {
            "description": "The thread",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ThreadId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThread"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningThread"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such thread"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "HARNESS_LOCKED: the thread already ran a turn on its harness, or is a fork; or COMMAND_CONFLICT"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "SETTING_NOT_OFFERED: a harness Shadows does not know"
          }
        },
        "summary": "Changes the thread's CLI before its first turn (spec §12.6). A replay\nanswers the thread as it now stands and changes nothing, even once the\nharness is locked; a new command after the first turn, or on a fork (its\nsession is a fork of its source's, §12.9), is `HARNESS_LOCKED`. A change\ncloses the thread's adapter (§12.2).",
        "tags": [
          "threads"
        ]
      }
    },
...
    "/api/threads/{id}/fork": {
      "post": {
        "operationId": "fork_thread",
        "parameters": [
          {
            "description": "The source thread",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ThreadId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForkThread"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningThread"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such thread or entry"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "THREAD_BUSY: a turn is running, or COMMAND_CONFLICT"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "FORK_POINT_NOT_SUPPORTED: fork from anything but the last completed entry"
          }
        },
        "summary": "Forks the thread from its last completed entry into a new thread on the\nsame project (spec §12.9). The new thread holds copies of the entries up to\nand including `at_entry_id`; its next turn continues a fork of the\nsource's harness session. The source is unchanged. A replay answers the\nfork already made.",
        "tags": [
          "threads"
        ]
      }
    },
...
  },
...
}